Purchase item

+++ server/WebAPI/Controllers/ItemController.cs
@@ -4,6 +4,7 @@ using Core.Models;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
+using MySqlX.XDevAPI.Common;
 using System.Security.Claims;
 using System.Threading.Tasks;
 
@@ -14,13 +15,15 @@ namespace WebAPI.Controllers
     public class ItemController : ControllerBase
     {
         private readonly IItemService _itemService;
+        private readonly IUserService _userService;
 
-        public ItemController(IItemService itemService)
+        public ItemController(IItemService itemService, IUserService userService)
         {
             _itemService = itemService;
+            _userService = userService;
         }
 
-        [Authorize(Roles = "Admin")]
+        [Authorize]
         [HttpGet]
         public async Task<IActionResult> GetAllItems()
         {
@@ -69,13 +72,29 @@ namespace WebAPI.Controllers
         [HttpDelete("{itemId:Guid}")]
         public async Task<IActionResult> DeleteItemById(Guid itemId)
         {
-            var result = await _itemService.DeleteAsync(itemId);
-            if (!result.Succeeded)
+            var itemRes = await _itemService.GetByIdAsync(itemId);
+            if(itemRes.Data == null)
             {
-                return NotFound(result);
+                return NotFound(itemRes);
             }
+            ClaimsIdentity? identity = HttpContext.User.Identity as ClaimsIdentity;
+            if(identity != null)
+            {
+                var userId = Guid.Parse(identity.FindFirst(ClaimTypes.NameIdentifier)!.Value);
+                if (itemRes.Data.User.Id.Equals(userId))
+                {
+                    return Forbid(nameof(DeleteItemById), userId.ToString());
+                }
+                
+                var result = await _itemService.DeleteAsync(itemId);
+                if (!result.Succeeded)
+                {
+                    return NotFound(result);
+                }
 
-            return Ok(result);
+                return Ok(result);
+            }
+            return Unauthorized();
         }
 
         [Authorize]
@@ -85,8 +104,15 @@ namespace WebAPI.Controllers
             ClaimsIdentity? identity = HttpContext.User.Identity as ClaimsIdentity;
             if(identity != null)
             {
-                var userId = identity.FindFirst(ClaimTypes.NameIdentifier)!.Value;
-                item.UserId = Guid.Parse(userId);
+                var userId = Guid.Parse(identity.FindFirst(ClaimTypes.NameIdentifier)!.Value);
+                item.UserId = userId;
+
+                var itemRes = await _itemService.GetByIdAsync(item.Id);
+                if (itemRes.Data.User.Id.Equals(userId))
+                {
+                    return Forbid(nameof(UpdateItem), userId.ToString());
+                }
+
                 var result = await _itemService.UpdateAsync(item);
                 if (!result.Succeeded)
                 {
@@ -101,15 +127,48 @@ namespace WebAPI.Controllers
 
         [Authorize]
         [HttpPatch]
-        public async Task<IActionResult> ChangeItemStatus([FromBody] ChangeItemStatusDto changeItemStatusDto)
+        public async Task<IActionResult> ChangeItemStatus(ChangeItemStatusDto changeItemStatusDto)
         {
-            var result = await _itemService.ChangeItemStatusAsync(changeItemStatusDto);
-            if (!result.Succeeded)
+            ClaimsIdentity? identity = HttpContext.User.Identity as ClaimsIdentity;
+            if(identity != null)
             {
-                return NotFound(result);
+                var userId = Guid.Parse(identity.FindFirst(ClaimTypes.NameIdentifier)!.Value);
+                var itemRes = await _itemService.GetByIdAsync(changeItemStatusDto.Id);
+
+                if (itemRes.Data.Id.Equals(userId))
+                {
+                    return Forbid(nameof(ChangeItemStatus), userId.ToString());
+                }
+
+                var result = await _itemService.ChangeItemStatusAsync(changeItemStatusDto);
+                if (!result.Succeeded)
+                {
+                    return NotFound(result);
+                }
+
+                return Ok(result);
             }
 
-            return Ok(result);
+            return Unauthorized();
+        }
+
+        [Authorize]
+        [HttpPost("purchase/{itemId:Guid}")]
+        public async Task<IActionResult> PurchaseItem(Guid itemId)
+        {
+            if (HttpContext.User.Identity is ClaimsIdentity identity)
+            {
+                Guid userId = Guid.Parse(identity.FindFirst(ClaimTypes.NameIdentifier)!.Value);
+                var result = await _itemService.PurchaseItemAsync(itemId, userId);
+                if(!result.Succeeded)
+                {
+                    return BadRequest(result);
+                }
+
+                return Ok(result);
+            }
+
+            return Unauthorized();
         }
     }
 }
